+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
- - Avoid spawning two new delimiters when instantiating a closed lambda function.
+ - Avoid spawning two scope opening nodes when instantiating a closed lambda function.
  - Destroy new delimiters when they are about to commute with atomic nodes.
 
 ## 0.5.0 - 2025-07-13