+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - Avoid spawning two scope opening nodes when instantiating a closed lambda function.
  - Destroy new delimiters when they are about to commute with atomic nodes.
+ - Immediately duplicate atomic nodes to optimize the graph size during reduction.
 
 ## 0.5.0 - 2025-07-13