+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
- - Statistics: measure the total number of allocated & freed nodes.
+ - Statistics: measure the total number of separately allocated & freed nodes.
 
 ## 0.6.0 - 2025-07-25